+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added loading alert for streamgage layer
 -   Added zoom level alert message to sidebar layer toggle
 -   Added WIM favicon
+-   Added Reference Point Icons for AQUARIUS layer
+-   Added zoom to AQUARIUS layer on map load
+-   Added popup to AQUARIUS layer
 
 ### Changed